requirements/mminstall.txt
mmcv-full==1.4.0
mmdet==2.14.0
mmsegmentation==0.14.1
# mmdet3d==0.16.0